Flutter application update
Added login, home page with todays bookings, service page with all services, and with the option to tap on a service, to show a service detail page. Then a Booking page with upcoming and finished bookings in separate tabs. Implemented so the service detail page, now shows available time slots for each service, that sorts by date.